added hi and lo pass filters to apu
as someone who hasn't taken a signals processing course, figuring
out how to get these things working was a real pain.

Luckilly, Wikipedia is aweomse, and straight-up provides an
algorithm for both lopass and hipass filters. All it needed was a
bit of modification, and bam, easy as that!

----

As a sidenote, as i've been developing the APU, i've compared it to
some of the other APUs out there (including Blarrg's), and it
definately sounds shittier.
Maybe it's because i don't have DMC implemented yet? Unlikely...

Odds are i'll have to look into actual waveform synthesis if I want
it to sound better, which will be fun-on-the-bun.

+++ src/nes/apu/apu.h
@@ -162,12 +162,11 @@ private:
 
   static const class AudioLUT {
   private:
-    short pulse_table [31];
-    short tnd_table  [203];
-
+    double pulse_table [31];
+    double tnd_table  [203];
   public:
     AudioLUT();
-    short sample(u8 pulse1, u8 pulse2, u8 triangle, u8 noise, u8 dmc) const;
+    double sample(u8 pulse1, u8 pulse2, u8 triangle, u8 noise, u8 dmc) const;
   } audioLUT;
 
 public:
